Ignore closed FastCGI stream.
With Apache Benchmark, it is possible to elicit an exception:

    ab -n 5000 -c 100 -t 60 'http://...'

Would result in:

    user error (FCGX_PutStr failed with error code: -1)

I have experienced this merely with a lighttpd and nginx FastCGI web site not even under particularly heavy load.

So it's a problem that this isn't handled. FCGX_PutStr returns an error if you try to write to a closed stream.

So we now handle that case by discontinuing the response as there is nothing to respond to, and going about our business.